consolidated-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
azure-functions==1.18.0
azure-communication-messages==1.0.0
azure-identity==1.15.0
azure-servicebus==7.11.4
azure-cosmos==4.5.1
openai==1.3.7
aiohttp==3.9.1
python-dotenv==1.0.0
pydantic==2.5.2
starlette-gzip-request==0.1.0
